+++ docs/UX Patterns/Table/Table.stories.tsx
@@ -1,200 +1,30 @@
-import type { ReactNode } from 'react'
-import { CopyOutlined } from '@ant-design/icons'
-import { faker } from '@faker-js/faker'
 import type { Meta, StoryObj } from '@storybook/react'
 import {
   Flex,
   Icon,
   Input,
-  Select,
-  Badge,
-  type IBadgeProps,
   Table,
-  type TableProps,
-  Tag,
-  type ITagProps,
-  Typography,
   Space,
-  Tooltip,
+  Button,
+  Checkbox,
+  Collapse,
+  ConfigProvider,
+  Divider,
+  type ICollapseProps,
+  Modal,
+  Select,
+  Typography,
+  TreeSelect,
+  Row,
+  Col,
 } from 'src/components'
 import { DatePickerWithDisabledYears } from 'src/components/data-entry/DatePicker/DatePicker.stories'
-import { ColorError, ColorSuccess, ColorTextPlaceholder } from 'src/styles/style'
-
-interface DataType {
-  key: string
-  name: string
-  id: string
-  timestamp: number
-  output: string
-  environment: Environment
-  status: Status
-  mpId: string
-}
-
-type Environment = 'development' | 'production'
-type Status = 'draft' | 'error' | 'ready'
-
-const EnvironmentColors: Record<Environment, ITagProps['color']> = {
-  production: 'blue',
-  development: 'purple',
-}
-
-const EnvironmentNames: Record<Environment, string> = {
-  production: 'Prod',
-  development: 'Dev',
-}
-
-const getTagColorForEnvironment = (env: Environment): ITagProps['color'] => EnvironmentColors[env]
-
-const getNameForEnvironment = (env: Environment) => EnvironmentNames[env]
-
-const StatusColors: Record<Status, IBadgeProps['color']> = {
-  draft: ColorTextPlaceholder,
-  error: ColorError,
-  ready: ColorSuccess,
-}
-
-const StatusNames: Record<Status, string> = {
-  draft: 'Draft',
-  error: 'Error',
-  ready: 'Ready',
-}
-
-const getStatusColor = (status: Status) => StatusColors[status]
-
-const getStatusName = (status: Status) => StatusNames[status]
-
-const columns: TableProps<DataType>['columns'] = [
-  {
-    title: 'Name',
-    dataIndex: 'name',
-    key: 'name',
-    render: (name: string) => {
-      const path = window.location.pathname.split('/')
-      path.pop()
-      const route = `${path.join('/')}/${name}`
-
-      return <Typography.Link href={route}>{name}</Typography.Link>
-    },
-  },
-  {
-    title: () => (
-      <Flex align="center" gap={2}>
-        <Typography.Text>ID</Typography.Text>
-        <Tooltip
-          title={
-            <>
-              <Typography.Text style={{ color: 'white' }}>Help lorem ipsum. </Typography.Text>
-              <Typography.Link href="/" style={{ color: 'white', textDecoration: 'underline' }}>
-                Learn More
-              </Typography.Link>
-            </>
-          }>
-          <Icon name="help" size="sm" />
-        </Tooltip>
-      </Flex>
-    ),
-    dataIndex: 'id',
-    key: 'id',
-  },
-  {
-    title: 'Timestamp (UTC)',
-    dataIndex: 'timestamp',
-    key: 'timestamp',
-    render: (timestampInMicroseconds: number): string => {
-      return new Date(timestampInMicroseconds / (1000 * 1000)).toLocaleString(undefined, {
-        month: 'short',
-        day: '2-digit',
-        year: 'numeric',
-        hour: 'numeric',
-        minute: '2-digit',
-        second: '2-digit',
-        timeZone: 'UTC',
-        hour12: false,
-      })
-    },
-  },
-  {
-    title: 'mPID',
-    dataIndex: 'mpId',
-    key: 'mpId',
-    render: (mpId: string): ReactNode => {
-      return <Typography.Text copyable={{ icon: <CopyOutlined aria-label="Copy mPID" /> }}>{mpId}</Typography.Text>
-    },
-  },
-  {
-    title: 'Output',
-    dataIndex: 'output',
-    key: 'output',
-  },
-  {
-    title: 'Environment',
-    key: 'environment',
-    dataIndex: 'environment',
-    render: (env: Environment): React.ReactNode => {
-      return <Tag color={getTagColorForEnvironment(env)}>{getNameForEnvironment(env)}</Tag>
-    },
-  },
-  {
-    title: 'Status',
-    dataIndex: 'status',
-    key: 'status',
-    render: (status: Status): React.ReactNode => <Badge color={getStatusColor(status)} text={getStatusName(status)} />,
-  },
-  {
-    title: 'Actions',
-    dataIndex: 'actions',
-    key: 'actions',
-    render: (): ReactNode => (
-      <Select
-        suffixIcon={<Icon name="moreActions" />}
-        variant="borderless"
-        dropdownStyle={{ width: '200px' }}
-        value={null}
-        options={[
-          { label: 'Option 1', value: 'option1' },
-          {
-            label: (
-              <Tooltip title="Explaining of why this is disabled" placement="right">
-                <span>Option 2</span>
-              </Tooltip>
-            ),
-            value: 'option2',
-            disabled: true,
-          },
-          {
-            label: <span style={{ color: ColorError }}>Delete</span>,
-            value: 'option2',
-          },
-        ]}
-      />
-    ),
-  },
-]
-
-function createMockRow(): DataType {
-  return {
-    id: `JNBSK-${faker.number.int({ min: 1000, max: 9999 }).toString()}`,
-    key: faker.string.uuid(),
-    name: faker.helpers.arrayElement([
-      'NBCU',
-      'Remarkable Foods',
-      'Lulo Bank',
-      'Shift',
-      'Marks and Spencer',
-      "Zaxby's",
-    ]),
-    timestamp: faker.date.recent().valueOf() * 1000 * 1000,
-    mpId: faker.number.int({ max: 9_999_999_999 }).toString(),
-    output: faker.helpers.arrayElement(['Braze', 'mP Analytics', 'Cortex', 'Applytics', 'Google Analytics']),
-    environment: faker.helpers.arrayElement(['development', 'production']),
-    status: faker.helpers.arrayElement(['draft', 'error', 'ready']),
-  }
-}
-
-const data: DataType[] = faker.helpers.multiple(createMockRow, {
-  count: 45,
-})
+import { tableColumns, tableData, type TableDataType } from './TableStoryUtils'
+import { SelectWithRangePicker } from 'docs/Candidate Components/Directory/Date Range Filter/SelectWithRangePicker'
+import { useState } from 'react'
+import { ColorTextDescription } from 'src/styles/style'
+import MinusSquareOutlined from '@ant-design/icons/MinusSquareOutlined'
+import PlusSquareOutlined from '@ant-design/icons/PlusSquareOutlined'
 
 const meta: Meta<typeof Table> = {
   title: 'UX Patterns/Table/Table',
@@ -207,8 +37,7 @@ export default meta
 
 type Story = StoryObj<typeof Table>
 
-export const BasicTable: Story = {
-  name: 'Basic Table',
+export const Primary: Story = {
   render: () => (
     <Space direction="vertical" style={{ width: '100%' }}>
       <Space direction="vertical" style={{ width: '100%' }}>
@@ -224,7 +53,292 @@ export const BasicTable: Story = {
           />
         </Flex>
       </Space>
-      <Table<DataType> columns={columns} dataSource={data} scroll={{ x: 'max-content' }} />
+      <Table<TableDataType> columns={tableColumns} dataSource={tableData} scroll={{ x: 'max-content' }} />
     </Space>
   ),
+}
+
+const TIME_OPTIONS = [
+  {
+    value: 'last12hours',
+    label: 'Last 12 hours',
+  } as const,
+  {
+    value: 'last7days',
+    label: 'Last 7 days',
+  } as const,
+  {
+    value: 'last14days',
+    label: 'Last 14 days',
+  } as const,
+]
+
+const FormSectionHeader = ({ label }: { label: string }) => (
+  <Typography.Text
+    style={{
+      marginBottom: '10px',
+      display: 'block',
+      color: ColorTextDescription,
+    }}
+    size="sm"
+    strong>
+    {label}
+  </Typography.Text>
+)
+
+interface IMultipleCheckboxItem {
+  name: string
+  value: unknown
+}
+
+interface IMultipleCheckboxesProps {
+  label: string
+  fieldName: string
+  items: IMultipleCheckboxItem[]
+}
+
+type ICollapsibleFormSectionProps = NonNullable<ICollapseProps['items']>[0]
+
+const CollapsibleSection = ({ ...item }: ICollapsibleFormSectionProps): React.JSX.Element => {
+  return (
+    <ConfigProvider
+      theme={{
+        components: {
+          Collapse: {
+            headerPadding: '2px 0',
+            contentPadding: '0 24px',
+          },
+        },
+      }}>
+      <Collapse
+        ghost
+        expandIcon={({ isActive }): React.JSX.Element => (isActive ? <MinusSquareOutlined /> : <PlusSquareOutlined />)}
+        items={[{ ...item, key: 'item' }]}
+      />
+    </ConfigProvider>
+  )
+}
+
+const MultipleCheckboxes = ({ label, items }: IMultipleCheckboxesProps) => (
+  <CollapsibleSection
+    label={label}
+    extra={
+      <Typography.Text style={{ cursor: 'text' }} disabled>
+        1 selected
+      </Typography.Text>
+    }>
+    <Space direction="vertical">
+      {items.map(({ name, value }) => (
+        <Checkbox key={name}>{name}</Checkbox>
+      ))}
+    </Space>
+  </CollapsibleSection>
+)
+
+function useModal(): [() => void, React.ReactElement] {
+  const [open, setOpen] = useState(false)
+
+  const context = (
+    <Modal
+      open={open}
+      destroyOnClose
+      onCancel={() => {
+        setOpen(false)
+      }}
+      cancelText={'Cancel'}
+      onOk={() => setOpen(false)}
+      okText={'Apply'}
+      width={'580px'}
+      style={{ maxHeight: '80vh' }}
+      closable
+      title={
+        <Typography.Text strong={false} size="xl">
+          Sort & Filters
+        </Typography.Text>
+      }>
+      <Flex vertical style={{ overflow: 'auto', maxHeight: '80vh' }}>
+        <FormSectionHeader label="Sorting" />
+        <Row>
+          <Col span={12}>
+            <Typography.Text>Order</Typography.Text>
+          </Col>
+          <Col span={12}>
+            <Select style={{ width: '100%' }}>
+              <Select.Option value={'false'}>Recent first</Select.Option>
+              <Select.Option value={'true'}>Oldest first</Select.Option>
+            </Select>
+          </Col>
+        </Row>
+        <Divider style={{ margin: '4px 0' }} />
+        <CollapsibleSection label="Name">
+          <Select
+            showSearch
+            style={{ width: '100%' }}
+            placeholder="Select name"
+            filterOption={(input, option) =>
+              typeof option?.label === 'string' && option.label.toLowerCase().includes(input.toLowerCase())
+            }
+            options={[
+              { value: 'NBCU', label: 'NBCU' },
+              { value: 'Remarkable Foods', label: 'Remarkable Foods' },
+              { value: 'Lulo Bank', label: 'Lulo Bank' },
+              { value: 'Shift', label: 'Shift' },
+              { value: 'Marks and Spencer', label: 'Marks and Spencer' },
+              { value: "Zaxby's", label: "Zaxby's" },
+            ]}
+          />
+        </CollapsibleSection>
+        <Divider style={{ margin: '4px 0' }} />
+        <CollapsibleSection label="Environment">
+          <Select
+            style={{ width: '100%' }}
+            options={[
+              { value: 'development', label: 'Development' },
+              { value: 'production', label: 'Production' },
+            ]}
+          />
+        </CollapsibleSection>
+        <Divider style={{ margin: '4px 0' }} />
+        <MultipleCheckboxes
+          fieldName="statuses"
+          label={'Status'}
+          items={[
+            { name: 'Draft', value: 'draft' },
+            { name: 'Error', value: 'error' },
+            { name: 'Production', value: 'production' },
+          ]}
+        />
+        <Divider style={{ margin: '4px 0' }} />
+        <CollapsibleSection label={'mPID'}>
+          <Input placeholder="mPID" allowClear />
+        </CollapsibleSection>
+        <Divider style={{ margin: '4px 0' }} />
+        <CollapsibleSection label={'Inputs'}>
+          <Select
+            mode="tags"
+            style={{ width: '100%' }}
+            options={[
+              { value: 'Braze', label: 'Braze' },
+              { value: 'mP Analytics', label: 'mP Analytics' },
+              { value: 'Cortex', label: 'Cortex' },
+              { value: 'Applytics', label: 'Applytics' },
+              { value: 'Google Analytics', label: 'Google Analytics' },
+            ]}
+          />
+        </CollapsibleSection>
+        <Divider style={{ margin: '4px 0' }} />
+        <CollapsibleSection label={'Outputs'}>
+          <TreeSelect
+            style={{ width: '100%' }}
+            treeCheckable
+            treeData={[
+              {
+                value: 'Braze',
+                title: 'Braze',
+                key: 'Braze',
+                children: [
+                  {
+                    value: 'Braze-1',
+                    title: 'Braze 1',
+                    key: 'Braze-1',
+                    children: [
+                      { value: 'Braze-1-1', title: 'Braze 1-1', key: 'Braze-1-1' },
+                      { value: 'Braze-1-2', title: 'Braze 1-2', key: 'Braze-1-2' },
+                      { value: 'Braze-1-3', title: 'Braze 1-3', key: 'Braze-1-3' },
+                    ],
+                  },
+                ],
+              },
+              {
+                value: 'mP Analytics',
+                title: 'mP Analytics',
+                key: 'mP Analytics',
+                children: [
+                  {
+                    value: 'mP Analytics-1',
+                    title: 'mP Analytics 1',
+                    key: 'mP Analytics-1',
+                    children: [
+                      { value: 'mP Analytics-1-1', title: 'mP Analytics 1-1', key: 'mP Analytics-1-1' },
+                      { value: 'mP Analytics-1-2', title: 'mP Analytics 1-2', key: 'mP Analytics-1-2' },
+                      { value: 'mP Analytics-1-3', title: 'mP Analytics 1-3', key: 'mP Analytics-1-3' },
+                    ],
+                  },
+                ],
+              },
+              { value: 'Cortex', title: 'Cortex', key: 'Cortex' },
+              { value: 'Applytics', title: 'Applytics', key: 'Applytics' },
+              { value: 'Google Analytics', title: 'Google Analytics', key: 'Google Analytics' },
+            ]}
+          />
+        </CollapsibleSection>
+      </Flex>
+    </Modal>
+  )
+
+  const showModal = (): void => {
+    setOpen(true)
+  }
+
+  return [showModal, context] as const
+}
+
+export const WithComplexFilters: Story = {
+  name: 'Complex',
+  render: () => {
+    const [showModal, context] = useModal()
+
+    return (
+      <>
+        {context}
+        <Space direction="vertical" style={{ width: '100%' }}>
+          <Space direction="vertical" style={{ width: '100%' }}>
+            <Flex align={'center'} justify={'space-between'}>
+              <Flex gap={10}>
+                <SelectWithRangePicker
+                  value={'last14days'}
+                  placeholder={'Choose Time'}
+                  options={TIME_OPTIONS}
+                  formatOptions={{
+                    dateStyle: 'short',
+                    timeStyle: 'short',
+                    hour12: false,
+                  }}
+                  dropdownStyle={{ minWidth: 400 }}
+                  rangePickerProps={{
+                    showTime: true,
+                    showHour: true,
+                    showMinute: true,
+                    showSecond: false,
+                    disabledDate: (antdDayJS: any) => {
+                      const fourteenDaysInMs = 14 * 24 * 60 * 60 * 1000
+                      return antdDayJS.isBefore(new Date(Date.now() - fourteenDaysInMs))
+                    },
+                  }}
+                />
+                <Button onClick={() => showModal()}>
+                  <Flex align="center" gap={5}>
+                    <Icon color="inherit" name={'filter'} size="sm" />
+                    <span>Sort & Filters</span>
+                    <Icon color="inherit" name={'dropdownOpen'} size="sm" />
+                  </Flex>
+                </Button>
+              </Flex>
+              <Input
+                allowClear
+                prefix={<Icon size="sm" color="brand" name="search" />}
+                placeholder="Search"
+                style={{ width: '240px' }}
+              />
+            </Flex>
+          </Space>
+          <Table<TableDataType>
+            columns={tableColumns}
+            dataSource={tableData.slice(0, 2)}
+            scroll={{ x: 'max-content' }}
+          />
+        </Space>
+      </>
+    )
+  },
 }