+++ src/components/index.ts
@@ -4,7 +4,12 @@ export { Icon, type IIconProps } from './general/Icon/Icon'
 export { Rate, type IRateProps } from './not-prod-ready/Rate/Rate'
 export { Form, type IFormProps, type FormInstance } from './data-entry/Form/Form'
 export { TreeSelect, type ITreeSelectProps } from './data-entry/TreeSelect/TreeSelect'
-export { Select, type ISelectProps, type DefaultOptionType } from './data-entry/Select/Select'
+export {
+  Select,
+  type ISelectProps,
+  type SelectDefaultOptionType,
+  type SelectBaseOptionType,
+} from './data-entry/Select/Select'
 export { Mentions, type IMentionsProps } from './not-prod-ready/Mentions/Mentions'
 export { Radio, type IRadioProps } from './data-entry/Radio/Radio'
 export { ColorPicker, type IColorPickerProps } from './not-prod-ready/ColorPicker/ColorPicker'
@@ -116,19 +121,17 @@ export { SuiteLogo } from './navigation/GlobalNavigation/SuiteLogo'
 export { Typography } from './general/Typography/Typography'
 
 // UPS
-export { UserPreferencesService } from '../services/user-preferences/user-preferences'
+export { UserPreferencesService } from '../services/user-preferences'
 export { CompositeUserPreferencesService } from '../services/user-preferences/composite-user-preferences-service'
-export { type CompositeUserPreferences } from '../services/user-preferences/models/user-preferences/composite-user-preferences'
 export {
   UserPreferenceScopeType,
   type UserPreferenceDefinition,
   type UserPreferenceDefinitions,
+  type UserPreferencesPerScope,
 } from '../services/user-preferences/models/definitions'
 export {
-  type UserPreferences,
   USER_PREFERENCE_SCOPE_SEPARATOR,
   UserPreferenceGlobalScope,
-  type UserPreference,
   type UserPreferenceScope,
 } from '../services/user-preferences/models/storage-models'
 export {