+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## [1.36.1](https://github.com/mParticle/aquarium/compare/v1.36.0...v1.36.1) (2024-12-06)
+
+### Bug Fixes
+
+- rename notification icons ([#508](https://github.com/mParticle/aquarium/issues/508)) ([98ea06f](https://github.com/mParticle/aquarium/commit/98ea06f1c0f66d7ca469f77158b1d4cf75cb8f6f))
+
+# [1.36.0](https://github.com/mParticle/aquarium/compare/v1.35.0...v1.36.0) (2024-11-28)
+
+### Bug Fixes
+
+- typography component updates ([#497](https://github.com/mParticle/aquarium/issues/497)) ([d32d390](https://github.com/mParticle/aquarium/commit/d32d3900c3d89406a0f88bccd967ed356e8edfe5))
+
+### Features
+
+- add abSplit icon ([#499](https://github.com/mParticle/aquarium/issues/499)) ([8776f1c](https://github.com/mParticle/aquarium/commit/8776f1c2fced9cc75af92fb74960510b4e56833a))
+- add notificationSubscriptions icon ([#505](https://github.com/mParticle/aquarium/issues/505)) ([b4976bc](https://github.com/mParticle/aquarium/commit/b4976bca3878813fadb453ff2426154b7fa1ef04))
+- utilize generics in UserPreferencesService ([#504](https://github.com/mParticle/aquarium/issues/504)) ([4838a86](https://github.com/mParticle/aquarium/commit/4838a86929bd743f606688ceefeaf13ebf5b1f9d))
+
+# [1.35.0](https://github.com/mParticle/aquarium/compare/v1.34.0...v1.35.0) (2024-11-15)
+
+### Bug Fixes
+
+- (styles) update font weight of modal ([#484](https://github.com/mParticle/aquarium/issues/484)) ([c2842ac](https://github.com/mParticle/aquarium/commit/c2842ac5494d2ec22e9bee9c0ff980922a67e2f4))
+- update release env variable ([#483](https://github.com/mParticle/aquarium/issues/483)) ([180c043](https://github.com/mParticle/aquarium/commit/180c0437c66578de3f0e8af25741f708c76c2845))
+
+### Features
+
+- DragAndDrop icon ([#492](https://github.com/mParticle/aquarium/issues/492)) ([fd2872d](https://github.com/mParticle/aquarium/commit/fd2872d7f19c50bd438da216685d4edf07efee71))
+
 # [1.34.0](https://github.com/mParticle/aquarium/compare/v1.33.2...v1.34.0) (2024-10-29)
 
 ### Features